[CDF-22072] 👻Splitting resource loaders into multiple files (#758)
* refactor; split reosurec loders into multiple files

* tests: updated test
