+++ cognite_toolkit/_cdf_tk/constants.py
@@ -26,6 +26,10 @@ ALT_CUSTOM_MODULES = "modules"
 ROOT_MODULES = [COGNITE_MODULES, CUSTOM_MODULES, ALT_CUSTOM_MODULES]
 MODULE_PATH_SEP = "/"
 
+MIN_TIMESTAMP_MS = -2208988800000  # 1900-01-01 00:00:00.000
+MAX_TIMESTAMP_MS = 4102444799999  # 2099-12-31 23:59:59.999
+HAS_DATA_FILTER_LIMIT = 10
+
 # Add any other files below that should be included in a build
 EXCL_FILES = ["README.md", DEFAULT_CONFIG_FILE]
 # Files to search for variables.
