+++ tests/tests_unit/utils.py
@@ -81,9 +81,21 @@ def mock_read_yaml_file(
     monkeypatch.setattr(
         "cognite_toolkit._cdf_tk.loaders._base_loaders.load_yaml_inject_variables", fake_load_yaml_inject_variables
     )
-    monkeypatch.setattr(
-        "cognite_toolkit._cdf_tk.loaders._resource_loaders.load_yaml_inject_variables", fake_load_yaml_inject_variables
-    )
+    for module in [
+        "auth_loaders",
+        "data_organization_loaders",
+        "datamodel_loaders",
+        "extraction_pipeline_loaders",
+        "file_loader",
+        "function_loaders",
+        "timeseries_loaders",
+        "transformation_loaders",
+        "workflow_loaders",
+    ]:
+        monkeypatch.setattr(
+            f"cognite_toolkit._cdf_tk.loaders._resource_loaders.{module}.load_yaml_inject_variables",
+            fake_load_yaml_inject_variables,
+        )
 
 
 class PrintCapture:
